+++ src/main/java/edu/ie3/simona/api/data/results/ontology/ResultDataMessageFromExt.java
@@ -1,0 +1,10 @@
+/*
+ * © 2024. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.simona.api.data.results.ontology;
+
+/** Messages that are sent from an external simulation to the SIMONA */
+public interface ResultDataMessageFromExt {}